app/requirements.txt
bcrypt
Flask==2.3.0
Flask-Migrate
Flask-Bcrypt
Flask-Login
Flask-SQLAlchemy
Flask-WTF
psycopg2-binary
Jinja2
SQLAlchemy==1.3.23
WTForms
celery
python-dotenv
